refactor: 기존 컴포넌트구조 변경
