+++ src/pages/GatheringDetail/components/GatheringGuide/index.tsx
