+++ src/pages/GatheringDetail/index.tsx
@@ -1,0 +1,14 @@
+import TabBar from '@/components/TabBar';
+
+const GatheringDetailPage = () => {
+  return (
+    <TabBar.Container>
+      <TabBar.Left>
+        <TabBar.GoBackButton />
+        <TabBar.Title>{}</TabBar.Title>
+      </TabBar.Left>
+    </TabBar.Container>
+  );
+};
+
+export default GatheringDetailPage;
